Remove partition size validation as with expiry policy partition sizes may differ. Move expire to under busyLock as pendingTree should be updated together with dataTree entry.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/GridCacheOffheapManager.java
@@ -791,13 +791,21 @@ public class GridCacheOffheapManager extends IgniteCacheOffheapManagerImpl imple
         if (!hasPendingEntries)
             return false;
 
-        int cleared = 0;
+        if (!busyLock.enterBusy())
+            return false;
 
-        for (CacheDataStore store : cacheDataStores()) {
-            cleared += ((GridCacheDataStore)store).purgeExpired(cctx, c, amount - cleared);
+        try {
+            int cleared = 0;
 
-            if (amount != -1 && cleared >= amount)
-                return true;
+            for (CacheDataStore store : cacheDataStores()) {
+                cleared += ((GridCacheDataStore)store).purgeExpired(cctx, c, amount - cleared);
+
+                if (amount != -1 && cleared >= amount)
+                    return true;
+            }
+        }
+        finally {
+            busyLock.leaveBusy();
         }
 
         return false;
